Moved out methods related to VM remote access to a separate mixin
