+++ app/controllers/vm_controller.rb
@@ -3,7 +3,8 @@ class VmController < ApplicationController
   before_action :get_session_data
   after_action :cleanup_action
   after_action :set_session_data
-  include VmCommon        # common methods for vm controllers
+  include VmCommon # common methods for vm controllers
+  include VmRemote # methods for VM remote access
 
   def index
     session[:vm_type] = nil             # Reset VM type if coming in from All tab
